+++ algorithms/dfs/dfs.hpp
@@ -4,5 +4,5 @@
 std::vector<bool> &dfs(
     const std::vector<std::vector<int64_t>> &graph,
     std::vector<bool> &reachable,
-    int node
+    int64_t node
 );
